rename, bump svelte2tsx
